separated param constants
